[Pdm2-9] Extract components from home to better atomize it
