+++ src/app/index.tsx
@@ -1,18 +1,6 @@
-import { View } from "react-native";
-import Background from "../common/components/Background";
-import Watermark from "../common/components/Watermark";
 import HomeScreen from "./Home/screens/HomeScreen";
 import React from "react";
 
 export default function index() {
-  return (
-    <Background>
-      <Watermark />
-      <View className="h-full justify-center items-center">
-        <Background.Transparent>
-          <HomeScreen />
-        </Background.Transparent>
-      </View>
-    </Background>
-  );
+  return <HomeScreen />;
 }
